build(nix): latest package versions with flake
Now using the latest packages versions with flakes. Previously the
latest versions of the packages (scikit-learn etc) were not working
with the 2D implementation. The isses was with PyMKS. PyMKS cutoffs
for two point correlations now work differently from the version being
used in shell.nix. The cutoff has been reduced from 30 to 25 to work
correctly.